requirements.txt
python-dotenv
pytz
python-telegram-bot
aiogram>=3.13,<4
asyncmy>=0.2
APScheduler>=3.10
python-dotenv>=1.0
pytz>=2023.3
tzlocal>=5.2
